better files for python API (not everything inside __init__.py(i) files
