--- apis/python/fastformat/converter/__init__.pyi
@@ -1,1 +1,0 @@
-
