+++ apis/python/fastformat/converter/arrow/arrow.pyi
