+++ apis/python/fastformat/datatypes/image.pyi
@@ -39,25 +39,3 @@ class Image:
 
     def name(self) -> str:
         ...
-
-class BBox:
-    @staticmethod
-    def new_xyxy(data: np.ndarray, confidence: np.ndarray, label: List[str]) -> BBox:
-        ...
-
-    @staticmethod
-    def new_xywh(data: np.ndarray, confidence: np.ndarray, label: List[str]) -> BBox:
-        ...
-
-    def into_xyxy(self) -> Self:
-        ...
-
-    def into_xywh(self) -> Self:
-        ...
-
-    def into_arrow(self) -> pa.UnionArray:
-        ...
-
-    @staticmethod
-    def from_arrow(array_data: pa.UnionArray) -> BBox:
-        ...
